RESTAPI- add response entity

+++ src/main/java/com/springrest/springrest/services/CourseServiceimpl.java
@@ -1,26 +1,22 @@
 package com.springrest.springrest.services;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import com.springrest.springrest.dao.CourseDao;
 import com.springrest.springrest.entity.Course;
+import com.springrest.springrest.exception.CourseNotFoundException;
 
 @Service
 public class CourseServiceimpl implements CourseService {
 	
-//	List<Course>list;
 	@Autowired
 	private CourseDao courseDao;
 	
-	public CourseServiceimpl()
-	{
-//		list = new ArrayList<>();
-//		list.add(new Course(125,"Java Core","Basic Java"));
-//		list.add(new Course(126,"Rest API","Create Rest API"));
 
-	}
 
 	@Override
 	public List<Course> getCourses() {
@@ -29,37 +25,27 @@ public class CourseServiceimpl implements CourseService {
 	}
 
 	@Override
-	public Course getCourse(long courseId) {
+	public Course getCourseById(long courseId) {
+		Course course=  courseDao.findById(courseId).orElse(null);
+		 if(course==null)
+				throw new CourseNotFoundException("course id "+courseId+" not present");
+		 return course;
 		
-//		Course c = null;
-//		for(Course course:list)
-//		{
-//			if(course.getId()==courseId)
-//			{
-//				c = course;
-//				break;
-//			}
-//		}
-		return courseDao.getOne(courseId);
+		
+			
+
 	}
 
 	@Override
 	public Course addCourse(Course course) {
 		
-//		list.add(course);
 		courseDao.save(course);
 		return course;
 	}
 
 	@Override
 	public Course updateCourse(Course course) {
-//		list.forEach(e ->{
-//			if (e.getId()==course.getId())
-//			{
-//				e.setTitle(course.getTitle());
-//				e.setDescription(course.getDescription());
-//			}
-//		});
+
 		courseDao.save(course);
 		return course;
 	}
@@ -67,7 +53,9 @@ public class CourseServiceimpl implements CourseService {
 	@Override
 	public void deleteCourse(long parseLong) 
 	{
-	Course entity = courseDao.getOne(parseLong);
+	Course entity = courseDao.findById(parseLong).orElse(null);
+	if(entity==null)
+		throw new CourseNotFoundException("course id "+parseLong+" not present");
 	courseDao.delete(entity);
 	}
